chore: FIT-339: Remove feature flags ff_front_dev_2715_audio_3_280722_short,ff_front_DEV_1713_audio_ui_150222_short
