+++ label_studio/feature_flags.json
@@ -433,33 +433,7 @@
       "version": 4,
       "deleted": false
     },
-    "ff_front_DEV_1713_audio_ui_150222_short": {
-      "key": "ff_front_DEV_1713_audio_ui_150222_short",
-      "on": false,
-      "prerequisites": [],
-      "targets": [],
-      "contextTargets": [],
-      "rules": [],
-      "fallthrough": {
-        "variation": 0
-      },
-      "offVariation": 1,
-      "variations": [
-        true,
-        false
-      ],
-      "clientSideAvailability": {
-        "usingMobileKey": false,
-        "usingEnvironmentId": false
-      },
-      "clientSide": false,
-      "salt": "b9a9a797e1884581b1ab307952adbb2c",
-      "trackEvents": false,
-      "trackEventsFallthrough": false,
-      "debugEventsUntilDate": null,
-      "version": 2,
-      "deleted": false
-    },
+
     "ff_front_dev_1442_unselect_shape_on_click_outside_080622_short": {
       "key": "ff_front_dev_1442_unselect_shape_on_click_outside_080622_short",
       "on": false,
@@ -649,33 +623,7 @@
       "version": 2,
       "deleted": false
     },
-    "ff_front_dev_2715_audio_3_280722_short": {
-      "key": "ff_front_dev_2715_audio_3_280722_short",
-      "on": true,
-      "prerequisites": [],
-      "targets": [],
-      "contextTargets": [],
-      "rules": [],
-      "fallthrough": {
-        "variation": 0
-      },
-      "offVariation": 1,
-      "variations": [
-        true,
-        false
-      ],
-      "clientSideAvailability": {
-        "usingMobileKey": false,
-        "usingEnvironmentId": false
-      },
-      "clientSide": false,
-      "salt": "a8eb8122a118496eab36bacd9dffff6b",
-      "trackEvents": false,
-      "trackEventsFallthrough": false,
-      "debugEventsUntilDate": null,
-      "version": 4,
-      "deleted": false
-    },
+
     "ff_front_optic_1494_saved_templates_to_custom_templates": {
       "key": "ff_front_optic_1494_saved_templates_to_custom_templates",
       "on": true,
