+++ web/libs/editor/src/components/SidePanels/DetailsPanel/RegionEditor.tsx
@@ -15,7 +15,6 @@ import {
 import { IconPropertyAngle } from "@humansignal/icons";
 import { Checkbox, Select } from "@humansignal/ui";
 import { Block, Elem, useBEM } from "../../../utils/bem";
-import { FF_DEV_2715, isFF } from "../../../utils/feature-flags";
 import { TimeDurationControl } from "../../TimeDurationControl/TimeDurationControl";
 import { TimelineRegionEditor } from "./TimelineRegionEditor";
 import "./RegionEditor.scss";
@@ -53,7 +52,7 @@ const IconMapping = {
 };
 
 const RegionEditorComponent: FC<RegionEditorProps> = ({ region }) => {
-  const isAudioRegion = isFF(FF_DEV_2715) && region.type === "audioregion";
+  const isAudioRegion = region.type === "audioregion";
   const isTimelineRegion = region.type === "timelineregion";
   const Component = isTimelineRegion ? TimelineRegionEditor : isAudioRegion ? AudioRegionProperties : RegionProperties;
