--- web/libs/editor/src/components/Waveform/Waveform.module.scss
@@ -1,16 +1,0 @@
-.progress {
-  color: var(--color-accent-persimmon-bold);
-}
-
-.wave {
-  position: relative;
-
-  canvas {
-    // prevent reset.css from breaking waveforms
-    max-width: unset;
-  }
-}
-
-.menu {
-  margin: 2em 0;
-}
