+++ web/libs/editor/src/core/feature-flags/flags.json
@@ -1,6 +1,4 @@
 {
-  "ff_front_DEV_1713_audio_ui_150222_short": false,
-  "ff_front_dev_2715_audio_3_280722_short": true,
   "fflag_feat_front_dev_3873_labeling_ui_improvements_short": true,
   "fflag_fix_front_lsdv_4620_memory_leaks_100723_short": false
 }
