+++ web/libs/editor/src/regions/AudioRegion.js
@@ -4,9 +4,7 @@ import NormalizationMixin from "../mixins/Normalization";
 import RegionsMixin from "../mixins/Regions";
 import { AreaMixin } from "../mixins/AreaMixin";
 import Registry from "../core/Registry";
-import { FF_DEV_2715, isFF } from "../utils/feature-flags";
 
-import { AudioUltraRegionModel as _audioUltraRegionModel } from "./AudioRegion/AudioUltraRegionModel";
 import { AudioRegionModel as _audioRegionModel } from "./AudioRegion/AudioRegionModel";
 import { EditableRegion } from "./EditableRegion";
 
@@ -37,7 +35,7 @@ const EditableAudioModel = types.model("EditableAudioModel", {}).volatile(() => 
   ],
 }));
 
-const AudioRegionModel = types.compose(
+const _exportAudioRegion = types.compose(
   "AudioRegionModel",
   RegionsMixin,
   AreaMixin,
@@ -46,22 +44,6 @@ const AudioRegionModel = types.compose(
   EditableAudioModel,
   _audioRegionModel,
 );
-
-const AudioUltraRegionModel = types.compose(
-  "AudioRegionModel",
-  RegionsMixin,
-  AreaMixin,
-  NormalizationMixin,
-  EditableRegion,
-  EditableAudioModel,
-  _audioUltraRegionModel,
-);
-
-let _exportAudioRegion = AudioRegionModel;
-
-if (isFF(FF_DEV_2715)) {
-  _exportAudioRegion = AudioUltraRegionModel;
-}
 
 Registry.addRegionType(_exportAudioRegion, "audioplus");
 Registry.addRegionType(_exportAudioRegion, "audio");
