+++ web/libs/editor/src/tags/object/Audio/view.tsx
@@ -24,14 +24,14 @@ const isAudioSpectrograms = isFF(FF_AUDIO_SPECTROGRAMS);
 
 const useSpectrogramControls = isAudioSpectrograms ? useSpectrogramControlsHook : () => {};
 
-interface AudioUltraProps {
+interface AudioProps {
   item: any;
   settings?: TimelineSettings;
   changeSetting?: (key: string, value: any) => void;
   children: ReactNode;
 }
 
-const AudioUltraView: FC<AudioUltraProps> = ({ item, children, settings = {}, changeSetting = () => {} }) => {
+const AudioView: FC<AudioProps> = ({ item, children, settings = {}, changeSetting = () => {} }) => {
   const rootRef = useRef<HTMLElement | null>();
   const isDarkMode = getCurrentTheme() === "Dark";
 
@@ -229,7 +229,7 @@ const AudioUltraView: FC<AudioUltraProps> = ({ item, children, settings = {}, ch
   );
 };
 
-const AudioUltraWithSettings: FC<AudioUltraProps> = ({ item }) => {
+const AudioWithSettings: FC<AudioProps> = ({ item }) => {
   const [settings, setSettings] = usePersistentJSONState<TimelineSettings>("ls:audio-tag:settings", {
     // @todo this hotkey should be moved from these settings for a more appropriate place;
     // @todo we are planning to have a central hotkeys management, that would be a better option.
@@ -261,13 +261,13 @@ const AudioUltraWithSettings: FC<AudioUltraProps> = ({ item }) => {
 
   return (
     <TimelineContextProvider value={contextValue}>
-      <AudioUltraView item={item}>
+      <AudioView item={item}>
         {item.errors?.map((error: any, i: number) => (
           <ErrorMessage key={`err-${i}`} error={error} />
         ))}
-      </AudioUltraView>
+      </AudioView>
     </TimelineContextProvider>
   );
 };
 
-export const AudioUltra = observer(AudioUltraWithSettings);
+export const Audio = observer(AudioWithSettings);
