--- web/libs/editor/src/tags/object/AudioPlus/AudioPlus.module.scss
@@ -1,9 +1,0 @@
-.play {
-  background: #52c41a;
-  border: 1px solid #52c41a;
-
-  &:hover {
-    background: #73d13d;
-    border: 1px solid #73d13d;
-  }
-}
