+++ web/libs/editor/tests/integration/feature-flags.ts
@@ -2,6 +2,5 @@ import * as FLAGS from "../../src/utils/feature-flags";
 
 export const CURRENT_FLAGS = {
   [FLAGS.FF_LSDV_4930]: true,
-  [FLAGS.FF_DEV_2715]: true,
   [FLAGS.FF_LSDV_4620_3_ML]: false,
 };
