+++ web/package.json
@@ -113,7 +113,6 @@
     "storybook": "^9.0.13",
     "strman": "^2.0.1",
     "tailwind-merge": "^2.6.0",
-    "wavesurfer.js": "^6.0.1",
     "webfft": "^1.0.3",
     "xpath-range": "^1.1.1",
     "zod": "^3.22.4"
